fix(Build): Fix Windows build
Currently, building Kaoto on Windows is broken due to a problem
resolving the base path of the catalog related to the folder separator.

The fix is to use the `pathToFileURL` function at all times when
resolving paths, so the folder and file paths gets properly resolved.

fix: https://github.com/KaotoIO/kaoto/issues/1592