Fix loading of Red Hat artifacts for catalog
io.kaoto.camelcatalog.maven.CamelCatalogVersionLoader.configureRepositories(String)
is called too late for Camel Yaml DSL loading. it is called when loading
Camel Catalogs.
Quickfix provided: calling the CamelCatalog which is configuring the
repositories at first
not perfect as note all versions are aligned and some others might need
it but should be enough to unblock current situation

note that the provided is not efficient because the camel-yaml-dsl test
dependency is leaking in the other tests, thus even without the fix the
test is passing... I guess it would require to create a separate module
for the test or to ensure a specific version is picked (see also next
paragraph)

Another note: when several catalog are generated at same time, I'm not
sure that we can ensure that it picks the exact correct version of the
camel-yaml-dsl, it might be the first one found, whatever version it is;
To be checked. In practice I think that most of the time there are no
differences.

fixes #1597

Signed-off-by: Aurélien Pupier <apupier@redhat.com>