feat(Canvas): Improve add new steps
Currently, we have 3 alternative mechanisms for adding new steps in the
canvas:

1. Right click on a node and select between `Prepend`, `Append`, and
   `Insert`.
2. Use the Step Toolbar to `Insert`
3. Use the quick add edge.

This commit replaces the quick add edge with a floating button over the
last element of the route. In addition to that, it generates a
placeholder when the steps property is empty, so the user can quickly
add a new step to initialize the branch.