fix(Build): Run Kaoto in Windows
Currently, dev mode is broken in Windows since the `vite.config.js` file
is not copying the catalog files properly. In addition to that, the
files path are generated using the underlying OS path separator.

The fix is to use the `normalizePath` function from `vite`, so it uses a
linux-like path to copy the files. From the Catalog generator part, it
forces to use the `/` as path separator, so it loads properly in the
browser.

fix: https://github.com/KaotoIO/kaoto/issues/1654