feat(Expression): 3rd iteration of ExpressionField
At this time, we're extracting the logic from the `OneOfField` component
so we can reuse it for the `ExpressionFieldInner` component.

The reason for this is toco place the `SchemaList` component inside of
the `ArrayFieldWrapper`, so the field looks more compact.

In addition to that, the children are only rendered if they are defined
or, in case of array children, if the array has one element or more.