fix(Document Export): Fix multiline property value in the table
Fixes: #2096

It replaces the line breaks to `&#10;` and also put `<pre><code>` when the markdown is rendered in the preview through the component mapper. On the other hand it preserves the XML tags as-is in the markdown source code for download to keep property value polution minimal. The markdown renderer would have to disable HTML to render XML tags in the property value.