feat(DataMapper): Support reading multiple mappings on a same collection target field
The first step for https://github.com/KaotoIO/kaoto/issues/1845, deserializing multiple mappings on a same collection target field. The UI part of adding multiple mappings is to come separately.
Introduced `FieldItemNodeData extends MappingNodeData` which represents the field node but it actually exists in the XSLT outcome, to be distinguished from the `TargetFieldNodeData` which represents the document fields that doesn't yet participate in any mappings. While document field has a consistent ID, `FieldItemNodeData` gets an unique ID for each mappings. This enables to distinguish the multiple field items for the same collection target field.

Also added `resetMappings` menu to the DataMapper debugger