feat(DataMapper): Allow multiple mappings to the collection field
Fixes: https://github.com/KaotoIO/kaoto/issues/1845

The 2nd and final step for supporting multiple mappings to the same collection field, adding the UI part of "Add Mapping". It also adds "Add Conditional Mapping" menu which allows to directly add a conditional mapping such as `for-each`, which is a short cut for "Add Mapping" and then "Wrap with for-each" for example.