fix(Form): Prevent onSubmit event from firing
Currently, when pressing `enter` in some form fields, the `onSubmit` event, making the UI to reload, and in case of VS Code, that leaves the user into a blank page.

The fix is to intercept the `onSubmit` event and stop the default
process.

fix: https://github.com/KaotoIO/kaoto/issues/2210