fix(Canvas): Avoid auto-fitting canvas
Currently, when adding or removing steps in the canvas, there's a process
of auto-fitting that makes the current zoom and position to be lost.

The fix is to remove the `setFitToScreenOnLayout` method call and
schedule a single autofit when loading the routes for the first time.

In addition to that, this commit removes the double rendering of the
`Canvas` component due to the `SourceCodeProvider` being initialized
with an empty string (`""`) and then updated from the `LocalStorage`
with an `useEffect` call in the `Shell` component.

fix: https://github.com/KaotoIO/kaoto/issues/2216