fix(Canvas): Keep existing visibility
Going to the source code editor and pasting a complete new route definition, when returning back to the design page, the route is hidden by default.

The fix is to keep the existing routes visibility if possible. This is
possible by checking the IDs from the VisualEntities and comparing them
against the VisibleFlows object.

In case a route is renamed, when the VisibleFlowsProvider is rerendered,
both the VisualEntities and the VisibleFlows object will match. On the
other hand, changing the entity type or pasting something entirely new
in the Source Code editor will try to keep existing flows visible,
otherwise it will try to at least show the first VisualEntity.

fix: https://github.com/KaotoIO/kaoto/issues/2214