feat(DataMapper): Select a mapping
Fixes: https://github.com/KaotoIO/kaoto/issues/2204

This enables to select and highlight a mapping by clicking a field or a mapping line. If a field is clicked, the mappings clicked field participates are highlighted, which includes source fields, lines and target fields. If a line is clicked, it behaves as if the target field connected to that line is clicked.