fix(EnumField): Show `EnumField` for enum props
Currently, the `SNMP` component has a `enum` property without setting
`type: enum` and this causes for Kaoto to render a `BeanField` instead.

Considering that waiting for the fix at the Camel catalog side will
invalidate previos Camel versions, a workaround is provided to lower the
sensitivity of the `enum` check and allow to show a `EnumField` for
schemas containing a `enum` property.

fix: https://github.com/KaotoIO/kaoto/issues/2565