chore: Remove innerRef warning when loading Datamapper
Currently, the `Truncate` component from Patternfly seems to be passing
a `ref` property into the underlying DOM nodes and this is causing an
error to be thrown in the browser's console.

A workaround for this situation is to remove the `Truncate` component
and replace it with a simple `span` DOM node instead.