chore(Storybook): Consolidate Forms stories
Currently, hand-made mocks are used in the `Forms` stories, causing that whenever there's an `API` change in the `IVisualizationNode`, the mocks need to be updated accordingly.

This commit consolidates the forms-related stories into a single file and uses a real Camel Route Visual entity to reduce the amount of mocks needed, positioning the stories closer to a runtime usage.

From now on, when a new EIP or component needs to be showcased through Storybook, it's as simple as adding a new step in the stub route and creating a new template for it.