fix(DataMapper): Replace useState + useRef with useRef
Currently, we're setting the reference to the mapping links area using a
`useState` hook to store it.

Since this rerender it's not necessary, we can skip the `useState` hook
and use a `useRef` directly instead, this simplifies the
`MappingLinksContext`'s API.

This is also a prerequisite of updating react to v19.